Update P6_01_Notebook.ipynb
BoVW / firsts steps: aggregation of descriptors of any size, many attempts to find a robust code to deal with that dimension issue.
BoW : tsne visualizations for tf_idf & lda & nmf, with 7 topics comparison. idea was to get the most important topic, but this only match for few categories with nmf.  kmeans to 7 clusters could help to set the right topics numbers?